🚨 fix lint errors

+++ contracts/smart-account/modules/BatchedSessionRouterModule.sol
@@ -1,6 +1,8 @@
 // SPDX-License-Identifier: MIT
 pragma solidity 0.8.17;
 
+/* solhint-disable function-max-lines */
+
 import {BaseAuthorizationModule, UserOperation} from "./BaseAuthorizationModule.sol";
 import {ISessionValidationModule} from "./SessionValidationModules/ISessionValidationModule.sol";
 import {ISessionKeyManager} from "../interfaces/ISessionKeyManager.sol";
@@ -81,7 +83,7 @@ contract BatchedSessionRouter is BaseAuthorizationModule {
 
         uint256 length = sessionData.length;
         // iterate over batched operations
-        for (uint i; i < length; ) {
+        for (uint256 i; i < length; ) {
             // validate the sessionKey
             // sessionKeyManager reverts if something wrong
             ISessionKeyManager(sessionKeyManager).validateSessionKey(
